Modded Monsters
- Added the functionality to create mods with new monsters.

In Libraries\Farmhand\API, a new folder was created, "Monsters". In this
folder is two new classes, Monster and MonsterInformation.
MonsterInformation contains data required for monsters, most of this
data usually loaded from the "Data\Monsters" XML. Some other data such
as the monster's name and texture is also stored here. The Monster class
is where MonsterInformation is registered from mods and stored for
injectors(There's becoming a lot of classes named Monster. It might not
be necessary to name this Monster?).

In Libraries\Farmhand\API\Generic, a new class was created called
"ItemChancePair". This class mirrors ItemQuantityPair in functionality,
except storing an item and a drop chance instead. This is used by
MonsterInformation in storing monster item drop information.

In Libraries\Farmhand\Content, two new content injectors were created,
MonsterInjector and MonsterLoader. MonsterInjector is an injector which
injects MonsterInformation into Data\Monsters, allowing it to be loaded
and parsed naturally by StardewValley.Monsters.Monster objects.
MonsterLoader loads new texture files as if from the Characters\Monsters
folder, allowing Monster objects to load those naturally. Both were
added to the ContentManager.

A new event was created, in MineShaftEvents. This event is currently not
usable for its intended purpose, because it requires a return value. It
is meant to hook the method in MineShaft which decides which monsters to
load for a given level. It's arguments are in
EventArgsGetMonsterForThisLevel. Once events are returnable, it can be
spruced up and made ready for actual mod use, although it may be better
to handle it internally and let mods register spawn chances with some
parameters through Farmhand.

A new second pass library was created, call "FarmhandCharacter". Monster
is extended from NPC, which is extended from Character, so this library
will hopefully eventually expand to have more overrides than just
Monster. This override allows monsters created in a mod to override more
behaviors in their monsters. The Monster class contained in
Libraries\FarmhandCharacter\Monster.cs is the class which is actually
extended in mods for use as a monster object.

- Added TestMonsterMod example mod

This mod shows how to create a mod which adds new monsters to the game.
It uses the new functionality added in this commit to Farmhand.