Crops Modding
- Added capability to mod in crops to Farmhand

Two new classes were added to Libraries\Farmhand\API\Crops, Crop and
CropInformation. CropInformation handles the data which will be injected
into Data\\Crops, and Crop holds the CropInformation which will be
injected.

A new injector called CropInjector was added to inject into Data\\Crops,
and was registered with the ContentManager

A few changes were made to the Items API and the TextureUtility.
ItemInformation.ToString() provides the category string, but not the
category negative number which is expected to follow. I'm not sure why
ItemCategory and ItemType are structured the way they are, but I think
ItemType is supposed to be the string, and ItemCategory is the number,
when being parsed into actual item categories. I didn't want to mess
with this too much until I could find out why it was implemented like
this, so I added a crude, simple debug to allow seeds to be properly
categorized, so they can be used as seeds. TextureUtility was changed
because it seemed to be specifically targeting Game1.objectSpriteSheet
when it was supposed to be targeting the passed spritesheet, which was
preventing me from patching a texture into the crops spritesheet.

- Added an example mod to show how to add new crops

My art sucks, it works though.

I had to give the player items differently than the RecipeTestMod does,
because trying to give them the seeds with the seed class type prevented
the item from being in the seed category, even when the category and
negative number combo were properly passed from ItemInformation. Just
using Farmhand.API.Player.Player.AddObject(Id) instead of
AddObject<type>() in order to use the generic StardewValley.Object class
allowed proper parsing.

- Fixed the FarmhandCharacter library documentation outputing to the
wrong directory