Tool modding
- Added capability to mod in tools

Tool and ToolInformation were added to the API. ToolInformation holds
data, and Tool registers new tools and injects the textures into the
spritesheet.

AddTool was added to Farmhand.API.Player which allows the adding of
tools to the player's inventory.

- Added example mod for tool modding

You wave it around and it gives you parsnips what else could you
possibly ask for in a tool?