Shop stock editing and bugfixes
- Added an example of editing shop stock in the TestCropMod example mod

Edited both pierre's and the hospital's shop stock to include new items.
Editing the shop stock is fairly simple through events, so making some
kind of API tool to do it isn't required, although it would be nice,
especially to have shop stock order sorting. I'm not sure where a tool
to register that kind of stuff would go, but I'll be thinking about it.

-Added OnGetDwarfShopStock event

This won't be very useful until returnable events are working, and I
didn't want to add all shop events until they are. For now the example
doesn't even use this.

- Fixed bug in TestToolMod where the sprite sheet index is lost upon
reloading

There's a virtual method in StardewValley.Tool that is called which
messes up the sprite sheet index based on the tool upgrade level.
Overriding this and just returning immediately solves the problem.