ContentManager Hook Override
- Added HookRedirectConstructorFromBase options to allow hooking of
constructors with specific arguments. This was neccessary because if
left unspecified, only the first listed constructor in the base class
could be hooked. In order to do this, I added a new field to the
HookRedirectConstructorFromBaseAttribute class, "Parameters", which is
then used by the patchers to specify exactly which constructor needs to
be hooked in any given method. The changes to CecilHelper and
CecilContext mainly involed moving over to a FullName check for the
constructor, rather than just checking Name. An additional method was
added to CecilContext to allow for this kind of checking, called
"GetMethodDefinitionFullName", which functions exactly like
"GetMethodDefinition", except it checks by FullName. Since MethodInfo
doesn't provide a FullName property, a custom FullName string had to be
built in CecilHelper's RedirectConstructorFromBase method. This was done
for the non-generic utilizing overload, this functionality will have to
be added for constructors involving generics later.

- Used new constructor hooking features to properly hook all
instantiations of StardewValley.LocalizedContentManager to instead
initialize Farmhand.Content.ContentManager. This effectively solves the
issue of all the scattered temporary content managers, since now
everything goes through Farmhand's overriding ContentManager.