Big Craftables Factories and other new goodies
- Used the new redirect constructor to method hook, a factory was set up
which assures that mods which use a type which extends BigCraftable gets
that object instantiated for its big craftables

- Added an override during the first pass called StardewObject, whos
base class is StardewValley.Object. This override does the standard job
of forcing virtual base and setting the lowest field level to protected,
to make way for the second pass library StardewObject to do it's job

- Created a class in the second pass library Game called StardewObject
which extends from StardewValley.Object with a slight name change to
StardewObject to avoid confusion. All overridable methods were listed in
this class in order to provide opportunities for documentation

- Moved the override for BigCraftables to the Game second pass library,
and changed its base class to StardewObject

- The BigCraftable static elements that allow registration of new
BigCraftable information and types, as well as hosting the factory,
stayed behind during the new BigCraftable move to the second pass
library

- Removed AddBigCraftable from Farmhand.API.Player.Player, as AddObject
does the exact same job

- Fixed Big Craftables example mod to use these new standards, and added
an example method override that displays a message in the debugger
whenever it is left clicked