Shop Stock Injection API
- Added functionality to the API in order to allow mods to quickly and
easily register items to be sold in shops. Mods can select what shop
they want to put items in, what item to put in, how much the item will
cost in the shop, how much stock is available in the shop, and can even
use a delegate to decide if the item is on sale, returning a boolean
value. The following changes all contributed to this addition

- Fixed RedirectConstructorFromBase hook parameters so that generic
parameters in the constructor get formatted as strings properly, in
CecilHelper

- Created an override of ShopMenu in the first pass libraries that
changes the base protection values of StardewValley.Menus.ShopMenu

- Created a class in the API called ShopUtilities that handles mod
requests for new shop stock. ShopUtilities remembers the injected stock
for every shop, and also handles requests from the ShopMenu second pass
override for information about what needs to be injected.

- Created an override of ShopMenu in the second pass libraries which
hooks every instantiated instance of ShopMenu. This override is what
actually figures out what shop it is, and then injects the proper items,
provided by ShopUtilities.

- Festival shops, and any other obscure ones I might have missed, are
not currently injectable options, this will have to be added later,
perhaps when more festival functionality has been added to the API.