xTile Map Merging
- Added the capability for the API to merge any number of conflicting
map files for the same location into a single map file, with warnings
about where changes conflict and overlap in multiple map files.

- Added the capability for mod manifests to register and load .tide and
.tbin map files for use in mods

- Added a MapInformation class to store information about maps for use
in mods

- Added a LocationUtilities class, where the actual merging of maps goes
down, as well as registering MapInformation objects to be used in
merging. This could/should be reworked a bit later, to have better
organization and also allow for easier registering of new locations and
so forth.

- Reworked the content injectors a little to be able to pass an output
by reference, if needed, since it wasn't working for injecting the maps

- Included an example mod with 2 map files, each editing the bus stop,
but making different edits. One adds a potted plant, the other a barrel.
Both are merged into the final map.