Map Merging Improvements
- Changed parameters to allow mods to independently request maps
provided to be merged and returned. It's not just for internal use
anymore!

- Added "SoftMerge" to the MapOverride enum, when a MapInformation
object is tagged with it, any tile conflicts it has with another map
will resolve to use the other map's changes, without throwing a warning

- Allowed MapInformation to specify an offset for the map to use,
displacing the map from 0,0. A warning will be displayed at a negative
offset, but it shouldn't cause outright errors. This will mainly come in
handy for small "sub-maps" to be merged together onto one bigger map.

- Created a warning that will be displayed if a MapInformation is
created that, between the size of the map and the offset used, would
create a map bigger than 155x155, which Entoarox says will crash the
game if used.

- Stored maps which have already been merged internally, so that they
don't have to be redundantly merged again later. This storage doesn't
apply to the maps merged for mods upon request.