Static Classes (#158)
* StardewObject ID mismatch fix

- Added a fix for StardewObjects to sort out their mismatched IDs when
reloadSprite() is called. A better implementation would be to do this
right after serialization for every object, but xml serialization
doesn't have that kind of functionality. This should help fix a few
things for now while we sort this out.

* ID Fixing

- Added functionality to fix ID conflicts after deserialization through
registering objects created with the default constructor with the proper
registry, then fixing all registered objects when deserialization is
complete.

- Changed and fix example mods and override classes to prevent use of
the default constructor for anything but being the default constructor.
Calling any base but the default base(base()) causes the object to not
be added to the registry, and not have its ID fixed.

- Objects only have their ID fixed if they are not the default type(such
as a StardewValley.Object). The ID checking relies on the object being a
child type of the default, so that the type is uniquely tied to the ID.
This means that if an item is created internally by StardewValley, and
it creates it as a StardewValley.Object, rather than an
ExampleMod.ExampleItem : StardewValley.Object, the ID fixer will have no
way of knowing what ID to fix it to. This can be fixed by the use of
factories, such as the one used already by
Farmhand.API.Items.BigCraftable.Create, which finds every instance of
instantiation of big craftable objects, and assures that if there is a
custom type registered by that ID, it is built using the type
registered. This should be done for everything using sequential IDs, to
assure that they are built as the correct type.

* Static Classes

- Made several registry classes static